graphrag-pattern-dashboard/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
matplotlib==3.7.2
google-generativeai==0.8.5
reportlab==4.0.4
networkx==3.1
psutil==5.9.5

# GraphRAG dependencies
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=1.0.6
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
pydantic>=2.0.0
chromadb>=0.4.0

# Database dependencies (optional - for production use)
psycopg2-binary==2.9.7
sqlalchemy==2.0.19
alembic==1.12.0